+++ android/src/main/java/com/ryanheise/just_audio/AudioPlayer.java
@@ -1,20 +1,38 @@
 package com.ryanheise.just_audio;
 
+import android.content.Context;
+import android.net.Uri;
 import android.os.Handler;
+
 import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.Player;
+import com.google.android.exoplayer2.ExoPlaybackException;
 import com.google.android.exoplayer2.PlaybackParameters;
+import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.SimpleExoPlayer;
+import com.google.android.exoplayer2.metadata.Metadata;
+import com.google.android.exoplayer2.metadata.MetadataOutput;
+import com.google.android.exoplayer2.metadata.icy.IcyHeaders;
+import com.google.android.exoplayer2.metadata.icy.IcyInfo;
 import com.google.android.exoplayer2.source.ClippingMediaSource;
 import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.source.ProgressiveMediaSource;
+import com.google.android.exoplayer2.source.TrackGroup;
+import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.source.dash.DashMediaSource;
 import com.google.android.exoplayer2.source.hls.HlsMediaSource;
+import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.upstream.DataSource;
 import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory;
 import com.google.android.exoplayer2.upstream.DefaultHttpDataSource;
+import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory;
 import com.google.android.exoplayer2.util.Util;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+import io.flutter.Log;
 import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.EventChannel.EventSink;
 import io.flutter.plugin.common.MethodCall;
@@ -22,16 +40,10 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.LinkedList;
-import android.content.Context;
-import android.net.Uri;
-import java.util.List;
 
-public class AudioPlayer implements MethodCallHandler, Player.EventListener {
+public class AudioPlayer implements MethodCallHandler, Player.EventListener, MetadataOutput {
+	static final String TAG = "AudioPlayer";
+
 	private final Registrar registrar;
 	private final Context context;
 	private final MethodChannel methodChannel;
@@ -55,18 +67,19 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 	private boolean buffering;
 	private boolean justConnected;
 	private MediaSource mediaSource;
+	private IcyInfo icyInfo;
+	private IcyHeaders icyHeaders;
 
 	private final SimpleExoPlayer player;
 	private final Handler handler = new Handler();
 	private final Runnable bufferWatcher = new Runnable() {
 		@Override
 		public void run() {
-			long newBufferedPosition = Math.min(duration, player.getBufferedPosition());
+			long newBufferedPosition = player.getBufferedPosition();
 			if (newBufferedPosition != bufferedPosition) {
 				bufferedPosition = newBufferedPosition;
 				broadcastPlaybackEvent();
 			}
-			if (duration > 0 && newBufferedPosition >= duration) return;
 			if (buffering) {
 				handler.postDelayed(this, 200);
 			} else if (state == PlaybackState.playing) {
@@ -100,6 +113,7 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 		state = PlaybackState.none;
 
 		player = new SimpleExoPlayer.Builder(context).build();
+		player.addMetadataOutput(this);
 		player.addListener(this);
 	}
 
@@ -109,15 +123,47 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 	}
 
 	@Override
+	public void onMetadata(Metadata metadata) {
+		for (int i = 0; i < metadata.length(); i++) {
+			final Metadata.Entry entry = metadata.get(i);
+			if (entry instanceof IcyInfo) {
+				icyInfo = (IcyInfo) entry;
+				broadcastPlaybackEvent();
+			}
+		}
+	}
+
+	@Override
+	public void onTracksChanged(TrackGroupArray trackGroups, TrackSelectionArray trackSelections) {
+		for (int i = 0; i < trackGroups.length; i++) {
+			TrackGroup trackGroup = trackGroups.get(i);
+
+			for (int j = 0; j < trackGroup.length; j++) {
+				Metadata metadata = trackGroup.getFormat(j).metadata;
+
+				if (metadata != null) {
+					for (int k = 0; k < metadata.length(); k++) {
+						final Metadata.Entry entry = metadata.get(k);
+						if (entry instanceof IcyHeaders) {
+							icyHeaders = (IcyHeaders) entry;
+							broadcastPlaybackEvent();
+						}
+					}
+				}
+			}
+		}
+	}
+
+	@Override
 	public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
 		switch (playbackState) {
 		case Player.STATE_READY:
 			if (prepareResult != null) {
-				duration = player.getDuration();
+				duration = getDuration();
 				justConnected = true;
+				transition(PlaybackState.stopped);
 				prepareResult.success(duration);
 				prepareResult = null;
-				transition(PlaybackState.stopped);
 			}
 			if (seekProcessed) {
 				completeSeek();
@@ -139,6 +185,27 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 				startWatchingBuffer();
 			}
 		}
+	}
+
+	@Override
+	public void onPlayerError(ExoPlaybackException error) {
+		switch (error.type) {
+		case ExoPlaybackException.TYPE_SOURCE:
+			Log.e(TAG, "TYPE_SOURCE: " + error.getSourceException().getMessage());
+			break;
+
+		case ExoPlaybackException.TYPE_RENDERER:
+			Log.e(TAG, "TYPE_RENDERER: " + error.getRendererException().getMessage());
+			break;
+
+		case ExoPlaybackException.TYPE_UNEXPECTED:
+			Log.e(TAG, "TYPE_UNEXPECTED: " + error.getUnexpectedException().getMessage());
+			break;
+
+		default:
+			Log.e(TAG, "default: " + error.getUnexpectedException().getMessage());
+		}
+		this.setError(String.valueOf(error.type), error.getMessage());
 	}
 
 	@Override
@@ -196,6 +263,9 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 				setSpeed((float)((double)((Double)args.get(0))));
 				result.success(null);
 				break;
+			case "setAutomaticallyWaitsToMinimizeStalling":
+				result.success(null);
+				break;
 			case "seek":
 				Object position = args.get(0);
 				if (position instanceof Integer) {
@@ -228,7 +298,39 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 		event.add(updatePosition = getCurrentPosition());
 		event.add(updateTime = System.currentTimeMillis());
 		event.add(Math.max(updatePosition, bufferedPosition));
-		eventSink.success(event);
+		event.add(collectIcyMetadata());
+		event.add(duration = getDuration());
+
+		if (eventSink != null) {
+			eventSink.success(event);
+		}
+	}
+
+	private ArrayList<Object> collectIcyMetadata() {
+		final ArrayList<Object> icyData = new ArrayList<>();
+		final ArrayList<String> info;
+		final ArrayList<Object> headers;
+		if (icyInfo != null) {
+			info = new ArrayList<>();
+			info.add(icyInfo.title);
+			info.add(icyInfo.url);
+		} else {
+			info = new ArrayList<>(Collections.nCopies(2, null));
+		}
+		if (icyHeaders != null) {
+			headers = new ArrayList<>();
+			headers.add(icyHeaders.bitrate);
+			headers.add(icyHeaders.genre);
+			headers.add(icyHeaders.name);
+			headers.add(icyHeaders.metadataInterval);
+			headers.add(icyHeaders.url);
+			headers.add(icyHeaders.isPublic);
+		} else {
+			headers = new ArrayList<>(Collections.nCopies(6, null));
+		}
+		icyData.add(info);
+		icyData.add(headers);
+		return icyData;
 	}
 
 	private long getCurrentPosition() {
@@ -238,6 +340,25 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 			return seekPos;
 		} else {
 			return player.getCurrentPosition();
+		}
+	}
+
+	private long getDuration() {
+		if (state == PlaybackState.none || state == PlaybackState.connecting) {
+			return C.TIME_UNSET;
+		} else {
+			return player.getDuration();
+		}
+	}
+
+	private void setError(String errorCode, String errorMsg) {
+		if (prepareResult != null) {
+			prepareResult.error(errorCode, errorMsg, null);
+			prepareResult = null;
+		}
+
+		if (eventSink != null) {
+			eventSink.error(errorCode, errorMsg, null);
 		}
 	}
 
@@ -256,19 +377,30 @@ public class AudioPlayer implements MethodCallHandler, Player.EventListener {
 		DataSource.Factory httpDataSourceFactory = new DefaultHttpDataSourceFactory(
 				userAgent,
 				DefaultHttpDataSource.DEFAULT_CONNECT_TIMEOUT_MILLIS,
-				DefaultHttpDataSource.DEFAULT_READ_TIMEOUT_MILLIS, 
+				DefaultHttpDataSource.DEFAULT_READ_TIMEOUT_MILLIS,
 				true
 		);
 		DataSource.Factory dataSourceFactory = new DefaultDataSourceFactory(context, httpDataSourceFactory);
 		Uri uri = Uri.parse(url);
-		if (uri.getPath().toLowerCase().endsWith(".mpd")) {
+		String extension = getLowerCaseExtension(uri);
+		if (extension.equals("mpd")) {
 			mediaSource = new DashMediaSource.Factory(dataSourceFactory).createMediaSource(uri);
-		} else if (uri.getPath().toLowerCase().endsWith(".m3u8")) {
+		} else if (extension.equals("m3u8")) {
 			mediaSource = new HlsMediaSource.Factory(dataSourceFactory).createMediaSource(uri);
 		} else {
 			mediaSource = new ProgressiveMediaSource.Factory(dataSourceFactory).createMediaSource(uri);
 		}
 		player.prepare(mediaSource);
+	}
+
+	private String getLowerCaseExtension(Uri uri) {
+		// Until ExoPlayer provides automatic detection of media source types, we
+		// rely on the file extension. When this is absent, as a temporary
+		// workaround we allow the app to supply a fake extension in the URL
+		// fragment. e.g.  https://somewhere.com/somestream?x=etc#.m3u8
+		String fragment = uri.getFragment();
+		String filename = fragment != null && fragment.contains(".") ? fragment : uri.getPath();
+		return filename.replaceAll("^.*\\.", "").toLowerCase();
 	}
 
 	public void setClip(final Long start, final Long end, final Result result) {